fix(subscriptions): avoid parallelSave from recursive fn call
