+++ website/server/libs/payments/google.js
@@ -9,6 +9,7 @@ import {
 import { model as IapPurchaseReceipt } from '../../models/iapPurchaseReceipt';
 import { model as User } from '../../models/user';
 import { getGemsBlock } from './gems';
+import promoSubscription from './promoSubscription'; // eslint-disable-line import/no-cycle
 
 const api = {};
 
@@ -218,6 +219,8 @@ api.noRenewSubscribe = async function noRenewSubscribe (options) {
     gift.subscription = sub;
     data.gift = gift;
     data.paymentMethod = this.constants.PAYMENT_METHOD_GIFT;
+
+    promoSubscription(data);
   }
 
   await payments.createSubscription(data);
